requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.0
mlflow>=1.20.0
dvc>=2.8.0
dvc-s3>=2.8.0
pytest>=6.2.5
pytest-cov>=2.12.1
python-dotenv>=0.19.0
monai>=0.8.0
dagshub>=0.1.7
